cheetah flutter

+++ binding/flutter/android/src/main/java/ai/picovoice/flutter/cheetah/CheetahPlugin.java
@@ -19,7 +19,12 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-import ai.picovoice.cheetah.*;
+import ai.picovoice.cheetah.Cheetah;
+import ai.picovoice.cheetah.CheetahException;
+import ai.picovoice.cheetah.CheetahInvalidArgumentException;
+import ai.picovoice.cheetah.CheetahInvalidStateException;
+import ai.picovoice.cheetah.CheetahRuntimeException;
+import ai.picovoice.cheetah.CheetahTranscript;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -28,16 +33,9 @@ import io.flutter.plugin.common.MethodChannel.Result;
 
 public class CheetahPlugin implements FlutterPlugin, MethodCallHandler {
 
-  private enum Method {
-    CREATE,
-    PROCESS,
-    FLUSH,
-    DELETE
-  }
-
+  private final Map<String, Cheetah> cheetahPool = new HashMap<>();
   private Context flutterContext;
   private MethodChannel channel;
-  private final Map<String, Cheetah> cheetahPool = new HashMap<>();
 
   @Override
   public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
@@ -65,10 +63,13 @@ public class CheetahPlugin implements FlutterPlugin, MethodCallHandler {
           String accessKey = call.argument("accessKey");
           String modelPath = call.argument("modelPath");
           double endpointDuration = call.argument("endpointDuration");
+          boolean enableAutomaticPunctuation = call.argument("enableAutomaticPunctuation");
 
-          Cheetah.Builder cheetahBuilder = new Cheetah.Builder().setAccessKey(accessKey)
+          Cheetah.Builder cheetahBuilder = new Cheetah.Builder()
+                  .setAccessKey(accessKey)
                   .setModelPath(modelPath)
-                  .setEndpointDuration((float) endpointDuration);
+                  .setEndpointDuration((float) endpointDuration)
+                  .setEnableAutomaticPunctuation(enableAutomaticPunctuation);
 
           Cheetah cheetah = cheetahBuilder.build(flutterContext);
           cheetahPool.put(String.valueOf(System.identityHashCode(cheetah)), cheetah);
@@ -87,27 +88,36 @@ public class CheetahPlugin implements FlutterPlugin, MethodCallHandler {
         }
         break;
       case PROCESS:
+
+        String handle = call.argument("handle");
+        ArrayList<Integer> pcmList = call.argument("frame");
+
+        if (!cheetahPool.containsKey(handle)) {
+          result.error(
+                  CheetahInvalidStateException.class.getSimpleName(),
+                  "Invalid cheetah handle provided to native module",
+                  null);
+          return;
+        }
+
+        Cheetah cheetah = cheetahPool.get(handle);
+        if (cheetah == null) {
+          result.error(
+                  CheetahInvalidStateException.class.getSimpleName(),
+                  "Instance of Cheetah no longer exists.",
+                  null);
+          return;
+        }
+
+        short[] pcm = null;
+        if (pcmList != null) {
+          pcm = new short[pcmList.size()];
+          for (int i = 0; i < pcmList.size(); i++) {
+            pcm[i] = pcmList.get(i).shortValue();
+          }
+        }
+
         try {
-          String handle = call.argument("handle");
-          ArrayList<Integer> pcmList = call.argument("frame");
-
-          if (!cheetahPool.containsKey(handle)) {
-            result.error(
-                    CheetahInvalidStateException.class.getSimpleName(),
-                    "Invalid cheetah handle provided to native module",
-                    null);
-            return;
-          }
-
-          short[] pcm = null;
-          if (pcmList != null) {
-            pcm = new short[pcmList.size()];
-            for (int i = 0; i < pcmList.size(); i++) {
-              pcm[i] = pcmList.get(i).shortValue();
-            }
-          }
-
-          Cheetah cheetah = cheetahPool.get(handle);
           CheetahTranscript transcriptObj = cheetah.process(pcm);
 
           Map<String, Object> param = new HashMap<>();
@@ -123,18 +133,27 @@ public class CheetahPlugin implements FlutterPlugin, MethodCallHandler {
         }
         break;
       case FLUSH:
+
+        String handle = call.argument("handle");
+
+        if (!cheetahPool.containsKey(handle)) {
+          result.error(
+                  CheetahInvalidStateException.class.getSimpleName(),
+                  "Invalid cheetah handle provided to native module",
+                  null);
+          return;
+        }
+
+        Cheetah cheetah = cheetahPool.get(handle);
+        if (cheetah == null) {
+          result.error(
+                  CheetahInvalidStateException.class.getSimpleName(),
+                  "Instance of Cheetah no longer exists.",
+                  null);
+          return;
+        }
+
         try {
-          String handle = call.argument("handle");
-
-          if (!cheetahPool.containsKey(handle)) {
-            result.error(
-                    CheetahInvalidStateException.class.getSimpleName(),
-                    "Invalid cheetah handle provided to native module",
-                    null);
-            return;
-          }
-
-          Cheetah cheetah = cheetahPool.get(handle);
           CheetahTranscript transcriptObj = cheetah.flush();
 
           Map<String, Object> param = new HashMap<>();
@@ -160,7 +179,9 @@ public class CheetahPlugin implements FlutterPlugin, MethodCallHandler {
         }
 
         Cheetah cheetah = cheetahPool.get(handle);
-        cheetah.delete();
+        if (cheetah != null) {
+          cheetah.delete();
+        }
         cheetahPool.remove(handle);
 
         result.success(null);
@@ -171,5 +192,12 @@ public class CheetahPlugin implements FlutterPlugin, MethodCallHandler {
   @Override
   public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     channel.setMethodCallHandler(null);
+  }
+
+  private enum Method {
+    CREATE,
+    PROCESS,
+    FLUSH,
+    DELETE
   }
 }
